% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Runs the parametric bootstrapping analysis for the best two models from the
empirical analysis of all candidate models.}
\usage{
bootstrap(
  data,
  archipelago,
  model_1,
  model_2,
  idparsopt_1,
  idparsopt_2,
  parsfix_1,
  parsfix_2,
  idparsfix_1,
  idparsfix_2,
  relaxed_model_1,
  relaxed_model_2,
  relaxed_par_1 = NULL,
  relaxed_par_2 = NULL,
  seed
)
}
\arguments{
\item{data}{list of DAISIE data}

\item{archipelago}{string with the name of the archipelago}

\item{model_1}{dataframe containing the maximum likelihood output of a model
from the function \code{archipelago_analysis}}

\item{model_2}{dataframe containing the maximum likelihood output of a model
from the function \code{archipelago_analysis}}

\item{idparsopt_1}{numeric vector with the id of parameters to optimise}

\item{idparsopt_2}{numeric vector with the id of parameters to optimise}

\item{parsfix_1}{numeric vector with the value of the fixed parameters}

\item{parsfix_2}{numeric vector with the value of the fixed parameters}

\item{idparsfix_1}{numeric vector with the id of parameters to fix}

\item{idparsfix_2}{numeric vector with the id of parameters to fix}

\item{relaxed_model_1}{boolean defining if the model is relaxed-rate}

\item{relaxed_model_2}{boolean defining if the model is relaxed-rate}

\item{relaxed_par_1}{string determining which parameter to relax, only required
if relaxed_model is TRUE
#' @param relaxed_par_2 string determining which parameter to relax, only required
if relaxed_model is TRUE}

\item{seed}{numeric seed for the simulation}
}
\value{
list of 4 elements each containing a dataframe with the maximum
likelihood parameter estimates from bootstrapping
}
\description{
Runs the parametric bootstrapping analysis for the best two models from the
empirical analysis of all candidate models.
}
\examples{
stub
}
