% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_lik_ratio.R
\name{bootstrap_lik_ratio}
\alias{bootstrap_lik_ratio}
\title{Title}
\usage{
bootstrap_lik_ratio(model_1, model_2, boostrap_output)
}
\arguments{
\item{boostrap_output}{}
}
\value{

}
\description{
Title
}
