% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_lik_ratio_alpha.R
\name{calc_lik_ratio_alpha}
\alias{calc_lik_ratio_alpha}
\title{Calculates the likelihood ratio significance value (lik_ratio_alpha) at
significance threshold of 0.05.}
\usage{
calc_lik_ratio_alpha(lik_ratio_model_1)
}
\arguments{
\item{lik_ratio_model_1}{numeric vector of likelihood ratios calculated from
fitting two models to a data set simulated under model 1}
}
\value{
numeric
}
\description{
Calculates the likelihood ratio significance value (lik_ratio_alpha) at
significance threshold of 0.05.
}
\examples{
output <- calc_lik_ratio_alpha(
lik_ratio_model_1 = runif(1000, 0, 100))
}
