% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_max_loglik.R
\name{calc_max_loglik}
\alias{calc_max_loglik}
\title{Finds the maximum loglikelihood and BIC value from a set of replicated models}
\usage{
calc_max_loglik(archipelago, model)
}
\arguments{
\item{archipelago}{string with the name of the archipelago}

\item{model}{string with the name of the model}
}
\value{
list of two numeric elements
}
\description{
Finds the maximum loglikelihood and BIC value from a set of replicated models
}
\examples{
output <- calc_max_loglik(archipelago = "Aldabra_Group",
model = "cr_di")
}
