% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_p_value.R
\name{calc_p_value}
\alias{calc_p_value}
\title{Calculates the p value for the bootstrapping likelihood ratio test from a
single value of likelihood ratio zeor (lik_ratio_0) and a set of likelihood
ratios (likelihood_ratio_model_1).}
\usage{
calc_p_value(lik_ratio_0, lik_ratio_model_1)
}
\arguments{
\item{lik_ratio_0}{numeric initial likelihood ratio calculated from the
likelihoods of the models fit to empirical data}

\item{lik_ratio_model_1}{numeric vector of likelihood ratios calculated from
fitting two models to a data set simulated under model 1}
}
\value{
numeric
}
\description{
Calculates the p value for the bootstrapping likelihood ratio test from a
single value of likelihood ratio zeor (lik_ratio_0) and a set of likelihood
ratios (likelihood_ratio_model_1).
}
\examples{
output <- calc_p_value(lik_ratio_0 = 10,
                                 lik_ratio_model_1 = runif(1000, 0, 100))
}
