% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_power.R
\name{calc_power}
\alias{calc_power}
\title{Calculates the power for the bootstrapping likelihood ratio test from a
single value of the likelihood ratio significance value (lik_ratio_alpha)
and a set of likelihood ratios (lik_ratio_model_2).}
\usage{
calc_power(lik_ratio_alpha, lik_ratio_model_2)
}
\arguments{
\item{lik_ratio_alpha}{The likelihood ratio associated with the significance
of the likelihood ratio test}

\item{lik_ratio_model_2}{numeric vector of likelihood ratios calculated from
fitting two models to a data set simulated under model 2}
}
\value{
numeric
}
\description{
Calculates the power for the bootstrapping likelihood ratio test from a
single value of the likelihood ratio significance value (lik_ratio_alpha)
and a set of likelihood ratios (lik_ratio_model_2).
}
