% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rr_approach_cr.R
\name{calc_rr_approach_cr}
\alias{calc_rr_approach_cr}
\title{Calculates the difference in likelihood for a clade (single lineage or
radiation) across different standard deviations of gamma distribution of
the relaxed-rate DAISIE model.}
\usage{
calc_rr_approach_cr(parameter = "cladogenesis", clade = "lineage")
}
\arguments{
\item{parameter}{A string defining which parameter to relaxed. Options are:
\code{"cladogenesis"}, \code{"extinction"}, \code{"carrying_capacity"}, or
\code{"anagenesis"}.}

\item{clade}{A string defining whether to use single lineage or a radiation.
Options are: \code{"lineage"} or \code{"radiation"}.}
}
\value{
A list of two numeric vectors
}
\description{
Calculates the difference in likelihood for a clade (single lineage or
radiation) across different standard deviations of gamma distribution of
the relaxed-rate DAISIE model.
}
