% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_sim_data.R
\docType{data}
\name{cr_sim_data}
\alias{cr_sim_data}
\title{Data simulated under the constant rate model}
\format{
An object of class \code{list}.
}
\usage{
cr_sim_data
}
\description{
A dataset containing simulated data for 48 parameter sets and 1000 replicates
per parameter set.
}
\keyword{datasets}
