% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_archipelagos.R
\name{create_archipelagos}
\alias{create_archipelagos}
\title{Extracts and saves individual archipelago data from the archipelagos41
dataset in the DAISIE R package}
\usage{
create_archipelagos()
}
\value{
String
}
\description{
Extracts and saves individual archipelago data from the archipelagos41
dataset in the DAISIE R package
}
\examples{
\dontrun{create_archipelagos()}
}
