% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_archipelagos_alt_m.R
\name{create_archipelagos_alt_m}
\alias{create_archipelagos_alt_m}
\title{Creates and saves archipelago data with alternative mainland species pool
size}
\usage{
create_archipelagos_alt_m()
}
\value{
String
}
\description{
Creates and saves archipelago data with alternative mainland species pool
size
}
\examples{
\dontrun{create_archipelagos_alt_m()}
}
