% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data.R
\name{create_data}
\alias{create_data}
\title{Extracts and saves individual archipelago data from the archipelagos41,
Bats_GreaterAntilles and frogs_datalist datasets in the DAISIE R package}
\usage{
create_data()
}
\value{
String
}
\description{
Extracts and saves individual archipelago data from the archipelagos41,
Bats_GreaterAntilles and frogs_datalist datasets in the DAISIE R package
}
