% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_max_loglik.R
\name{find_max_loglik}
\alias{find_max_loglik}
\title{Finds the maximum loglikelihood and BIC value from a set of replicated models}
\usage{
find_max_loglik(archipelago)
}
\arguments{
\item{archipelago}{string with the name of the archipelago}
}
\value{
list of two numeric elements
}
\description{
Finds the maximum loglikelihood and BIC value from a set of replicated models
}
