% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archipelago_analysis.R
\name{archipelago_analysis}
\alias{archipelago_analysis}
\title{Runs the maximum likelihood optimisation for a given DAISIE model}
\usage{
archipelago_analysis(
  data,
  archipelago,
  idparsopt,
  parsfix,
  idparsfix,
  ddmodel,
  seed,
  relaxed_model,
  relaxed_par = NULL,
  save_output = TRUE
)
}
\arguments{
\item{data}{list of DAISIE data}

\item{archipelago}{string with the name of the archipelago}

\item{idparsopt}{numeric vector with the id of parameters to optimise}

\item{parsfix}{numeric vector with the value of the fixed parameters}

\item{idparsfix}{numeric vector with the id of parameters to fix}

\item{ddmodel}{numeric determining diversity-independence or -dependence}

\item{seed}{numeric seed for the simulation}

\item{relaxed_model}{boolean defining if the model is relaxed-rate}

\item{relaxed_par}{string determining which parameter to relax, only required
if relaxed_model is TRUE}

\item{save_output}{boolean determining whether the output is saved or
returned}
}
\description{
Runs the maximum likelihood optimisation for a given DAISIE model
}
\examples{
output <- archipelago_analysis(
data = Aldabra_Group,
archipelago = "Aldabra_Group",
idparsopt = c(1,2,3,4,5),
parsfix = NULL,
idparsfix = NULL,
ddmodel = 0,
seed = 1,
relaxed_model = FALSE)
}
