% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_lik_ratio.R
\name{bootstrap_lik_ratio}
\alias{bootstrap_lik_ratio}
\title{Runs the parametric bootstrapping analysis for the best two models from the
empirical analysis of all candidate models.}
\usage{
bootstrap_lik_ratio(
  data,
  archipelago,
  model_1,
  model_2,
  idparsopt_1,
  idparsopt_2,
  parsfix_1,
  parsfix_2,
  idparsfix_1,
  idparsfix_2,
  ddmodel_1,
  ddmodel_2,
  relaxed_model_1,
  relaxed_model_2,
  relaxed_par_1 = NULL,
  relaxed_par_2 = NULL,
  seed,
  save_output = TRUE
)
}
\arguments{
\item{data}{list of DAISIE data}

\item{archipelago}{string with the name of the archipelago}

\item{model_1}{dataframe containing the maximum likelihood output of a model
from the function \code{archipelago_analysis}}

\item{model_2}{dataframe containing the maximum likelihood output of a model
from the function \code{archipelago_analysis}}

\item{idparsopt_1}{numeric vector with the id of parameters to optimise}

\item{idparsopt_2}{numeric vector with the id of parameters to optimise}

\item{parsfix_1}{numeric vector with the value of the fixed parameters}

\item{parsfix_2}{numeric vector with the value of the fixed parameters}

\item{idparsfix_1}{numeric vector with the id of parameters to fix}

\item{idparsfix_2}{numeric vector with the id of parameters to fix}

\item{ddmodel_1}{numeric determining diversity-dependence in the model}

\item{ddmodel_2}{numeric determining diversity-dependence in the model}

\item{relaxed_model_1}{boolean defining if the model is relaxed-rate}

\item{relaxed_model_2}{boolean defining if the model is relaxed-rate}

\item{relaxed_par_1}{string determining which parameter to relax, only
required if relaxed_model is TRUE}

\item{relaxed_par_2}{string determining which parameter to relax, only
required if relaxed_model is TRUE}

\item{seed}{numeric seed for the simulation}

\item{save_output}{boolean determining whether the output is saved or
returned}
}
\value{
list of 4 elements each containing a dataframe with the maximum
likelihood parameter estimates from bootstrapping
}
\description{
Runs the parametric bootstrapping analysis for the best two models from the
empirical analysis of all candidate models.
}
\examples{
output <- relaxedDAISIE::bootstrap_lik_ratio(
data = Aldabra_Group,
archipelago = "Aldabra_Group",
model_1 = data.frame(lambda_c = 0.8135786,
                     mu = 0.6530319,
                     K = 0.5210479,
                     gamma = 0.01645943,
                     lambda_a = 1.01026,
                     loglik = -73.36238,
                     df = 5,
                     conv = 0),
model_2 = data.frame(lambda_c = 0.8135786,
                     mu = 0.6530319,
                     K = 0.5210479,
                     gamma = 0.01645943,
                     lambda_a = 1.01026,
                     loglik = -73.36238,
                     df = 5,
                     conv = 0),
idparsopt_1 = c(1, 2, 3, 4, 5),
idparsopt_2 = c(1, 2, 4, 5),
parsfix_1 = NULL,
parsfix_2 = Inf,
idparsfix_1 = NULL,
idparsfix_2 = 3,
ddmodel_1 = 11,
ddmodel_2 = 0,
relaxed_model_1 = FALSE,
relaxed_model_2 = FALSE,
relaxed_par_1 = NULL,
relaxed_par_2 = NULL,
seed = 1,
save_output = FALSE)
}
