% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_lik_ratio.R
\name{bootstrap_lik_ratio}
\alias{bootstrap_lik_ratio}
\title{Runs the parametric bootstrapping analysis for the best two models from the
empirical analysis of all candidate models.}
\usage{
bootstrap_lik_ratio(
  data,
  archipelago,
  model_1,
  model_2,
  idparsopt_1,
  idparsopt_2,
  parsfix_1,
  parsfix_2,
  idparsfix_1,
  idparsfix_2,
  ddmodel_1,
  ddmodel_2,
  relaxed_model_1,
  relaxed_model_2,
  relaxed_par_1 = NULL,
  relaxed_par_2 = NULL,
  seed,
  save_output = TRUE
)
}
\arguments{
\item{data}{list of DAISIE data}

\item{archipelago}{string with the name of the archipelago}

\item{model_1}{dataframe containing the maximum likelihood output of a model
from the function \code{archipelago_analysis}}

\item{model_2}{dataframe containing the maximum likelihood output of a model
from the function \code{archipelago_analysis}}

\item{idparsopt_1}{numeric vector with the id of parameters to optimise}

\item{idparsopt_2}{numeric vector with the id of parameters to optimise}

\item{parsfix_1}{numeric vector with the value of the fixed parameters}

\item{parsfix_2}{numeric vector with the value of the fixed parameters}

\item{idparsfix_1}{numeric vector with the id of parameters to fix}

\item{idparsfix_2}{numeric vector with the id of parameters to fix}

\item{ddmodel_1}{numeric determining diversity-dependence in the model}

\item{ddmodel_2}{numeric determining diversity-dependence in the model}

\item{relaxed_model_1}{boolean defining if the model is relaxed-rate}

\item{relaxed_model_2}{boolean defining if the model is relaxed-rate}

\item{relaxed_par_1}{string determining which parameter to relax, only
required if relaxed_model is TRUE}

\item{relaxed_par_2}{string determining which parameter to relax, only
required if relaxed_model is TRUE}

\item{seed}{numeric seed for the simulation}

\item{save_output}{boolean determining whether the output is saved or
returned}
}
\value{
list of 4 elements each containing a dataframe with the maximum
likelihood parameter estimates from bootstrapping
}
\description{
Runs the parametric bootstrapping analysis for the best two models from the
empirical analysis of all candidate models.
}
