% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{Document to inherit the function documentation
 @param parameter String determining which parameter is relaxed
 @param clade String either "lineage" or "clade"}
\usage{
default_params_doc(parameter, clade)
}
\description{
Document to inherit the function documentation
 @param parameter String determining which parameter is relaxed
 @param clade String either "lineage" or "clade"
}
