% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{Document to inherit the function documentation}
\usage{
default_params_doc(
  data,
  archipelago,
  model,
  idparsopt,
  parsfix,
  idparsfix,
  ddmodel,
  seed,
  relaxed_model,
  relaxed_par,
  save_output,
  model_1,
  model_2,
  idparsopt_1,
  idparsopt_2,
  parsfix_1,
  parsfix_2,
  idparsfix_1,
  idparsfix_2,
  ddmodel_1,
  ddmodel_2,
  relaxed_model_1,
  relaxed_model_2,
  relaxed_par_1,
  relaxed_par_2,
  lik_ratio_0,
  lik_ratio_model_1,
  lik_ratio_model_2,
  lik_ratio_alpha,
  verbose
)
}
\arguments{
\item{data}{list of DAISIE data}

\item{archipelago}{string with the name of the archipelago}

\item{model}{string with the name of the model}

\item{idparsopt}{numeric vector with the id of parameters to optimise}

\item{parsfix}{numeric vector with the value of the fixed parameters}

\item{idparsfix}{numeric vector with the id of parameters to fix}

\item{ddmodel}{numeric determining diversity-independence or -dependence}

\item{seed}{numeric seed for the simulation}

\item{relaxed_model}{boolean defining if the model is relaxed-rate}

\item{relaxed_par}{string determining which parameter to relax, only required
if relaxed_model is TRUE}

\item{save_output}{boolean determining whether the output is saved or
returned}

\item{model_1}{dataframe containing the maximum likelihood output of a model
from the function \code{archipelago_analysis}}

\item{model_2}{dataframe containing the maximum likelihood output of a model
from the function \code{archipelago_analysis}}

\item{idparsopt_1}{numeric vector with the id of parameters to optimise}

\item{idparsopt_2}{numeric vector with the id of parameters to optimise}

\item{parsfix_1}{numeric vector with the value of the fixed parameters}

\item{parsfix_2}{numeric vector with the value of the fixed parameters}

\item{idparsfix_1}{numeric vector with the id of parameters to fix}

\item{idparsfix_2}{numeric vector with the id of parameters to fix}

\item{ddmodel_1}{numeric determining diversity-dependence in the model}

\item{ddmodel_2}{numeric determining diversity-dependence in the model}

\item{relaxed_model_1}{boolean defining if the model is relaxed-rate}

\item{relaxed_model_2}{boolean defining if the model is relaxed-rate}

\item{relaxed_par_1}{string determining which parameter to relax, only
required if relaxed_model is TRUE}

\item{relaxed_par_2}{string determining which parameter to relax, only
required if relaxed_model is TRUE}

\item{lik_ratio_0}{numeric initial likelihood ratio calculated from the
likelihoods of the models fit to empirical data}

\item{lik_ratio_model_1}{numeric vector of likelihood ratios calculated from
fitting two models to a data set simulated under model 1}

\item{lik_ratio_model_2}{numeric vector of likelihood ratios calculated from
fitting two models to a data set simulated under model 2}

\item{lik_ratio_alpha}{The likelihood ratio associated with the significance
of the likelihood ratio test}

\item{verbose}{logical determining whether function prints}
}
\description{
Document to inherit the function documentation
}
